user-service/requirements.txt
flask
python-dotenv
psycopg2
flask-bcrypt
flask-jwt-extended